+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tsvector column type
+- Functional Indexes
+- Column Index Options 
 
 ## [0.4.0] - 2021-09-18
 ### Added